requirements.txt
gspread~=6.1.3
google-auth
pandas~=2.2.3
scikit-learn~=1.5.2
numpy
